Mise à jour security0yaml,serviceController.php et des templates pour insertion d'un bouton de deconnexion

+++ src/Controller/Admin/ServiceController.php
@@ -10,7 +10,6 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Doctrine\Persistence\ManagerRegistry;
-use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/admin/service')]
 class ServiceController extends AbstractController
@@ -23,20 +22,24 @@ class ServiceController extends AbstractController
     }
 
     #[Route('/', name: 'admin_service_index', methods: ['GET'])]
-    #[IsGranted('ROLE_EMPLOYE')]
-    #[IsGranted('ROLE_ADMIN')]
     public function index(ServiceRepository $serviceRepository): Response
     {
+        if (!$this->isGranted('ROLE_EMPLOYE') && !$this->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException();
+        }
+
         return $this->render('admin/service/index.html.twig', [
             'services' => $serviceRepository->findAll(),
         ]);
     }
 
     #[Route('/new', name: 'admin_service_new', methods: ['GET', 'POST'])]
-    #[IsGranted('ROLE_EMPLOYE')]
-    #[IsGranted('ROLE_ADMIN')]
     public function new(Request $request): Response
     {
+        if (!$this->isGranted('ROLE_EMPLOYE') && !$this->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException();
+        }
+
         $service = new Service();
         $form = $this->createForm(ServiceType::class, $service);
         $form->handleRequest($request);
@@ -56,10 +59,12 @@ class ServiceController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'admin_service_edit', methods: ['GET', 'POST'])]
-    #[IsGranted('ROLE_EMPLOYE')]
-    #[IsGranted('ROLE_ADMIN')]
     public function edit(Request $request, Service $service): Response
     {
+        if (!$this->isGranted('ROLE_EMPLOYE') && !$this->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException();
+        }
+
         $form = $this->createForm(ServiceType::class, $service);
         $form->handleRequest($request);
 
@@ -79,6 +84,10 @@ class ServiceController extends AbstractController
     #[Route('/{id}', name: 'admin_service_delete', methods: ['POST'])]
     public function delete(Request $request, Service $service): Response
     {
+        if (!$this->isGranted('ROLE_EMPLOYE') && !$this->isGranted('ROLE_ADMIN')) {
+            throw $this->createAccessDeniedException();
+        }
+
         if ($this->isCsrfTokenValid('delete' . $service->getId(), $request->request->get('_token'))) {
             $entityManager = $this->doctrine->getManager();
             $entityManager->remove($service);
@@ -87,10 +96,11 @@ class ServiceController extends AbstractController
 
         return $this->redirectToRoute('admin_service_index');
     }
-
     #[Route('/logout', name: 'app_logout', methods: ['GET'])]
-    public function logout(): void
-    {
-        // Code vide car Symfony gère la déconnexion automatiquement
-    }
+public function logout(): void
+{
+    
+    //Je laisse le code vide ici car Symfony gère automatiquement la déconnexion
+}
+
 }
